Api Rest Integration

+++ Emtagas.Facturation/src/Emtagas.Facturacion.Core/Entities/DetalleFactura.cs
@@ -6,6 +6,6 @@ namespace Emtagas.Facturacion.Core.Entities
 
         public string Description { get; set; }
 
-        public int SubTotal { get; set; }
+        public decimal SubTotal { get; set; }
     }
 }
